Re-add input index delta compression. It turns out to complement grid-stride indexing nicely

+++ cuda_tp.cpp
@@ -6,7 +6,7 @@
 void tensor_product_forward_cuda(
     float *__restrict__ in1, float *__restrict__ in2, float *__restrict__ out,
     uint32_t *__restrict__ out_indices, float *__restrict__ cb_palette,
-    uint32_t *__restrict__ input_indices, size_t len_indices, size_t in1_size,
+    uint64_t *__restrict__ input_indices, size_t len_indices, size_t in1_size,
     size_t in2_size, size_t cb_palette_size, size_t out_size, int batch_size);
 
 torch::Tensor tensor_product_forward(torch::Tensor in1, torch::Tensor in2,
@@ -18,7 +18,7 @@ torch::Tensor tensor_product_forward(torch::Tensor in1, torch::Tensor in2,
   tensor_product_forward_cuda(
       in1.data<float>(), in2.data<float>(), out.data<float>(),
       out_indices.data<uint32_t>(), cb_palette.data<float>(),
-      input_indices.data<uint32_t>(), input_indices.size(0), in1.size(1),
+      input_indices.data<uint64_t>(), input_indices.size(0), in1.size(1),
       in2.size(1), cb_palette.size(0), out.size(1), in1.size(0));
 
   return out;
